make main swipable component in the Dashboard page, add new context for sharing info between pages
