+++ src/components/CryptoContext.tsx
@@ -1,20 +1,36 @@
 import { createContext, Dispatch, SetStateAction } from 'react';
 import { CryptoAction } from '../components/helpers/reducer/actions'
 export interface CryptoContextType {
-  allUserPairs: string[],
   userCurrentPair: string[],
   price: number,
   isGoingUp: boolean,
   dayChangePercentage: number
 }
 export const startInApp: CryptoContextType = {
-  allUserPairs: [],
   userCurrentPair: [],
   price: 0.00,
   isGoingUp: true,
   dayChangePercentage: 0
 }
 
+// ABOUT TO ADD A CONTEXT
+
+export interface PageContextType {
+  allUserPairs: string[],
+}
+
+export const startPage: PageContextType = {
+  allUserPairs: [],
+};
+
+export interface PageContextState {
+  pageContext: PageContextType,
+  setPageContext?: Dispatch<SetStateAction<PageContextType>>
+}
+
+const startPageContext: PageContextState = {
+  pageContext: startPage
+}
 export interface ContextState {
   context: CryptoContextType,
   setContext?: Dispatch<SetStateAction<CryptoContextType>>
@@ -36,5 +52,7 @@ const startDispatch: DispatchState = {
 export const CryptoContext = createContext(startContext);
 
 export const DispatchContext = createContext(startDispatch);
+
+export const PageContext = createContext(startPageContext);
 
 // <{context: CryptoContextType | null, setContext: Dispatch<SetStateAction<any>>} | null>(null);
