+++ src/pages/ChooseCrypto.tsx
@@ -15,7 +15,7 @@ import {
   Center,
   Box
 } from "@chakra-ui/react";
-import { CryptoContext, DispatchContext } from '../components/CryptoContext';
+import { CryptoContext, DispatchContext, PageContext } from '../components/CryptoContext';
 import router from 'next/router';
 import { SET_ALL_CHOSEN_PAIRS, SET_CURRENT_PAIR } from "../components/helpers/reducer/actions";
 
@@ -28,8 +28,8 @@ interface apiDataTypes {
 const ChooseCrypto:React.FC = () => {
 
   const [apiData, setApiData] = useState<apiDataTypes[]>([]);
-  const { context } = useContext(CryptoContext);
-  const { dispatch } = useContext(DispatchContext);
+  const { pageContext, setPageContext } = useContext(PageContext);
+  // const { dispatch } = useContext(DispatchContext);
 
 
 // Used solely to make unique Key for the children props for apiData.map()
@@ -51,29 +51,30 @@ const ChooseCrypto:React.FC = () => {
   const handleAddCrypto = (e:BaseSyntheticEvent) => {
     console.log(e.target.dataset.crypto);
     const chosenPair = e.target.dataset.crypto;
-    console.log(context);
-    if(context.allUserPairs.includes(chosenPair)){
+    // ? console.log(context);
+    if(pageContext.allUserPairs.includes(chosenPair)){
       console.log("already chosen!");
       return;
     }
     
-    // setContext!({...context, allUserPairs: context.allUserPairs.concat(chosenPair)});
-    dispatch({ 
-      type: SET_ALL_CHOSEN_PAIRS,
-      payload: context.allUserPairs.concat(chosenPair)
-     })
+    setPageContext!({...pageContext, allUserPairs: pageContext.allUserPairs.concat(chosenPair)});
+    // dispatch({ 
+    //   type: SET_ALL_CHOSEN_PAIRS,
+    //   payload: context.allUserPairs.concat(chosenPair)
+    //  })
   };
 
   const handleDone = (e:BaseSyntheticEvent) => {
     console.log("done");
-    // setContext!({
-    //   ...context, 
-    //   userCurrentPair: [context.allUserPairs[0]]
-    // });
-    dispatch({
-      type: SET_CURRENT_PAIR,
-      payload: [context.allUserPairs[0]]
-    })
+    // setPageContext!({
+    //   ...pageContext, 
+    //   userCurrentPair: [pageContext.allUserPairs[0]]
+    // })
+    // ! set userCurrentPair inside the Dashboard!
+    // dispatch({
+    //   type: SET_CURRENT_PAIR,
+    //   payload: [context.allUserPairs[0]]
+    // })
     router.push("/cryptoDashboard");
   }
