+++ src/pages/_app.tsx
@@ -1,16 +1,17 @@
 import { ChakraProvider } from '@chakra-ui/react'
 import Head from "next/head"
-import { useReducer } from 'react';
-import { ContextReducer } from '../components/helpers/reducer';
+import { useReducer, useState, useMemo } from 'react';
 import theme from '../theme';
 import { AppProps } from 'next/app';
-import { CryptoContext, DispatchContext, startInApp } from '../components/CryptoContext';
+import { CryptoContext, DispatchContext, startInApp, startPage, PageContext } from '../components/CryptoContext';
 
 function MyApp({ Component, pageProps }: AppProps) {
   // const [context, setContext] = useState(startInApp);
-  const [context, dispatch] = useReducer(ContextReducer, startInApp);
-  
+  // ! const [context, dispatch] = useReducer(ContextReducer, startInApp);
+  const [pageContext, setPageContext] = useState(startPage);
   // const providerValue = useMemo(() => ({ context, dispatch}), [context, dispatch])
+
+  const providerValue = useMemo(() => ({ pageContext, setPageContext}), [pageContext, setPageContext])
 
   return (
     <ChakraProvider resetCSS theme={theme}>
@@ -19,11 +20,13 @@ function MyApp({ Component, pageProps }: AppProps) {
           <meta name="keywords" content="Crypto-Watcher" ></meta>
       </Head>
       
-      <DispatchContext.Provider value={{ dispatch }}>
-        <CryptoContext.Provider value={{context}}>
-          <Component {...pageProps} />
-        </CryptoContext.Provider>
-      </DispatchContext.Provider>
+      {/* <DispatchContext.Provider value={{ dispatch }}> */}
+        {/* <CryptoContext.Provider value={{context}}> */}
+          <PageContext.Provider value={providerValue}>
+            <Component {...pageProps} />
+          </PageContext.Provider>
+        {/* </CryptoContext.Provider> */}
+      {/* </DispatchContext.Provider> */}
     </ChakraProvider>
   )
 }
